Move TimedGridWidget to widgets package
